Feat: convert step form to tab form
