+++ src/app/js/fields/wizard/TabIdentity.js
@@ -1,19 +1,14 @@
 import React from 'react';
-import translate from 'redux-polyglot/translate';
 import { Field, FieldArray } from 'redux-form';
 import PropTypes from 'prop-types';
 import { TextField as MUITextField } from '@material-ui/core';
 
-import Step from './Step';
 import FieldLanguageInput from '../FieldLanguageInput';
 import FieldLabelInput from '../FieldLabelInput';
 import FieldInternal from '../FieldInternal';
 import ClassList from '../ClassList';
 
-import {
-    polyglot as polyglotPropTypes,
-    field as fieldPropTypes,
-} from '../../propTypes';
+import { field as fieldPropTypes } from '../../propTypes';
 
 const TextField = ({
     label,
@@ -31,28 +26,30 @@ const TextField = ({
     />
 );
 
-export const StepIdentityComponent = ({
-    field,
-    isSubresourceField,
-    p: polyglot,
-    ...props
-}) => {
+export const TabIdentityComponent = ({ field }) => {
     return (
-        <Step id="step-identity" label="field_wizard_step_identity" {...props}>
+        <>
             <FieldLabelInput />
             <Field name="scope" component={TextField} type="hidden" />
             <FieldArray name="classes" component={ClassList} type="classes" />
             <FieldLanguageInput field={field} />
             <FieldInternal field={field} />
-        </Step>
+        </>
     );
 };
 
-StepIdentityComponent.propTypes = {
-    isSubresourceField: PropTypes.bool,
-    field: fieldPropTypes.isRequired,
-    p: polyglotPropTypes.isRequired,
-    filter: PropTypes.string,
+TextField.propTypes = {
+    label: PropTypes.string,
+    input: PropTypes.shape({ name: PropTypes.string }),
+    meta: PropTypes.shape({
+        touched: PropTypes.bool,
+        invalid: PropTypes.bool,
+        error: PropTypes.string,
+    }),
 };
 
-export default translate(StepIdentityComponent);
+TabIdentityComponent.propTypes = {
+    field: fieldPropTypes.isRequired,
+};
+
+export default TabIdentityComponent;
